refactor(*): readdress the enum class to other folders
